nodeRed_Python_servers/drill/requirements.txt
asyncua 
python-dotenv